Extracting VirusCheckService from module mixin
One of the major disadvantages of module mixins is that several methods
often get defined on the object. This greatly increases the chance of
method collisions if lots of mixins are used.

So, while preserving much of the outward interface of the
`Hydra::Works::VirusCheck`'s validation, I'm extracting a collaborating
class that can do the heavy lifting against the file.

This involved a bit of additional refactoring as I was encountering
challenges with ClamAV definitions. By relying also on dependency
injection of the system virus scanner, I can more easily test each
piece of the anti-virus process in better isolation.